+++ src/main/java/com/marklogic/debugger/web/ApiController.java
@@ -17,8 +17,8 @@ import org.apache.commons.io.IOUtils;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.http.client.ClientHttpRequest;
-import org.springframework.http.client.ClientHttpRequestFactory;
 import org.springframework.http.client.ClientHttpResponse;
 import org.springframework.http.client.SimpleClientHttpRequestFactory;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -150,14 +150,24 @@ public class ApiController {
 				return evalQuery(auth, "get-file.xqy", hm);
 		}
 
-		@RequestMapping(value = "/servers/{serverId}/attached", method = RequestMethod.GET)
+		@RequestMapping(value = "/servers/{serverId}/requests", method = RequestMethod.GET)
 		@ResponseBody
-		public String getAttached(@PathVariable String serverId) throws InvalidRequestException {
+		public String getRequests(@PathVariable String serverId) throws InvalidRequestException {
 			ConnectionAuthenticationToken auth = (ConnectionAuthenticationToken)SecurityContextHolder.getContext().getAuthentication();
 			HashMap<String, String> hm = new HashMap<>();
 			hm.put("serverId", serverId);
-			return evalQuery(auth, "get-attached.xqy", hm);
+			return evalQuery(auth, "get-requests.xqy", hm);
 		}
+
+	@RequestMapping(value = "/servers/{serverId}/invoke", method = RequestMethod.POST)
+	public ResponseEntity<?> invokeModule(@PathVariable String serverId, @RequestParam String uri) throws InvalidRequestException {
+		ConnectionAuthenticationToken auth = (ConnectionAuthenticationToken)SecurityContextHolder.getContext().getAuthentication();
+		HashMap<String, String> hm = new HashMap<>();
+		hm.put("serverId", serverId);
+		hm.put("uri", uri);
+		evalQuery(auth, "invoke-module.xqy", hm);
+		return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+	}
 
 	@RequestMapping(value = "/requests/{requestId}/stack", method = RequestMethod.GET)
 	@ResponseBody
@@ -204,13 +214,23 @@ public class ApiController {
 		return evalQuery(auth, "continue.xqy", hm);
 	}
 
+	@RequestMapping(value = "/requests/{requestId}/pause", method = RequestMethod.GET)
+	@ResponseBody
+	public String pauseRequest(@PathVariable String requestId) throws InvalidRequestException {
+		ConnectionAuthenticationToken auth = (ConnectionAuthenticationToken)SecurityContextHolder.getContext().getAuthentication();
+		HashMap<String, String> hm = new HashMap<>();
+		hm.put("requestId", requestId);
+		return evalQuery(auth, "pause.xqy", hm);
+	}
+
 	@RequestMapping(value = "/requests/{requestId}/breakpoints", method = RequestMethod.POST)
 	@ResponseBody
 	public String setBreakpoints(@PathVariable String requestId, @RequestBody List<Breakpoint> breakpoints) throws InvalidRequestException {
 		ConnectionAuthenticationToken auth = (ConnectionAuthenticationToken)SecurityContextHolder.getContext().getAuthentication();
+		HashMap<String, String> hm = new HashMap<>();
+		hm.put("requestId", requestId);
+		evalQuery(auth, "clear-breakpoints.xqy", hm);
 		for (Breakpoint bp : breakpoints) {
-			HashMap<String, String> hm = new HashMap<>();
-			hm.put("requestId", requestId);
 			hm.put("uri", bp.uri);
 			hm.put("line", bp.line);
 			evalQuery(auth, "set-breakpoints.xqy", hm);
@@ -224,7 +244,7 @@ public class ApiController {
 		ConnectionAuthenticationToken auth = (ConnectionAuthenticationToken)SecurityContextHolder.getContext().getAuthentication();
 		HashMap<String, String> hm = new HashMap<>();
 		hm.put("requestId", requestId);
-		return evalQuery(auth, "set-breakpoints.xqy", hm);
+		return evalQuery(auth, "get-breakpoints.xqy", hm);
 	}
 
 	@RequestMapping(value = "/requests/{requestId}/eval", method = RequestMethod.POST)